add cachestate to db

+++ java-app/src/main/java/uk/dioxic/mongotakeaway/generator/CustomerGenerator.java
@@ -1,5 +1,7 @@
 package uk.dioxic.mongotakeaway.generator;
 
+import com.mongodb.client.result.DeleteResult;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.context.event.EventListener;
@@ -14,13 +16,12 @@ import uk.dioxic.mongotakeaway.domain.Customer;
 import uk.dioxic.mongotakeaway.domain.GlobalProperties;
 import uk.dioxic.mongotakeaway.event.CustomersReloadedEvent;
 import uk.dioxic.mongotakeaway.event.PropertiesChangedEvent;
+import uk.dioxic.mongotakeaway.service.CacheStateService;
 import uk.dioxic.mongotakeaway.service.EventService;
 import uk.dioxic.mongotakeaway.service.GlobalLockService;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Locale;
-import java.util.Random;
+import javax.validation.constraints.NotNull;
+import java.util.*;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Function;
@@ -31,32 +32,23 @@ import static org.springframework.data.mongodb.core.query.Query.query;
 
 @Slf4j
 @Component
+@RequiredArgsConstructor
 public class CustomerGenerator {
-    //    private final static CountDownLatch dbInitialised = new CountDownLatch(1);
-    private final static Lock initialiseLock = new ReentrantLock();
     private final Random random = new Random();
     private final List<Customer> customers = new ArrayList<>();
-    private final ReactiveMongoTemplate mongoTemplate;
-    private final Faker faker;
+    private final Faker faker= Faker.instance(Locale.UK);
+
+    private final @NotNull ReactiveMongoTemplate mongoTemplate;
+    private final @NotNull CacheStateService cacheStateService;
+    private final @NotNull GlobalLockService lockService;
+    private final @NotNull EventService eventService;
+
     private GlobalProperties globalProperties;
-    private GlobalLockService lockService;
-    private EventService eventService;
 
-    public CustomerGenerator(ReactiveMongoTemplate mongoTemplate, GlobalLockService lockService, EventService eventService) {
-        this.mongoTemplate = mongoTemplate;
-        this.lockService = lockService;
-        this.eventService = eventService;
-        this.faker = Faker.instance(Locale.UK);
-    }
-
-    /**
-     * Blocks until the DB is initialised.
-     * <p/>
-     * Returns instantly if the DB has already been initialised.
-     */
-    private void initialiseDatabaseIfNeccessary() {
-        if (lockService.tryLock("CUSTOMER")) {
+    private boolean reload() {
+        if (cacheStateService.isDirty("CACHE") && lockService.tryLock("CUSTOMER")) {
             try {
+                log.info("reloading");
                 if (globalProperties.generator().isDropCollection()) {
                     log.info("dropping customer collection");
                     mongoTemplate.dropCollection(Customer.class)
@@ -74,10 +66,12 @@ public class CustomerGenerator {
                             .doOnSuccess(res -> log.info("purged all customers"))
                             .block();
                 } else {
-                    mongoTemplate.find(new Query().with(Sort.by("_id")).skip(globalProperties.generator().getCustomers()-1).limit(1), Customer.class)
+                    mongoTemplate.find(new Query().with(Sort.by("id")).skip(globalProperties.generator().getCustomers()-1).limit(1), Customer.class)
                             .doOnError(e -> log.error(e.getMessage(), e))
                             .flatMap(customer -> mongoTemplate.remove(query(where("id").gt(customer.getId())), Customer.class))
-                            .doOnNext(res -> log.info("purged {} superfluous customers", res.getDeletedCount()))
+                            .map(DeleteResult::getDeletedCount)
+                            .filter(count -> count > 0)
+                            .doOnNext(count -> log.info("purged {} superfluous customers", count))
                             .then(mongoTemplate.count(new Query(), Customer.class))
                             .map(count -> globalProperties.generator().getCustomers() - count)
                             .filter(count -> count > 0)
@@ -93,11 +87,14 @@ public class CustomerGenerator {
                             )
                             .block();
                 }
+                cacheStateService.markDirty("CACHE", false);
                 eventService.publishEvent(new CustomersReloadedEvent("reload"));
+                return true;
             } finally {
                 lockService.unlock("CUSTOMER");
             }
         }
+        return false;
     }
 
     private static Customer generateCustomer(Faker faker) {
@@ -131,22 +128,28 @@ public class CustomerGenerator {
         }
     }
 
-    @Async
-    @EventListener
-    public void handlePropertiesChanged(PropertiesChangedEvent event) {
-        globalProperties = event.getSource();
-        initialiseDatabaseIfNeccessary();
-    }
-
-    @Async
-    @EventListener
-    public void handleCustomersReloaded(CustomersReloadedEvent event) {
+    private void refresh() {
         mongoTemplate.find(new Query().limit(globalProperties.generator().getCustomers()), Customer.class)
                 .doOnError(e -> log.error(e.getMessage(), e))
                 .collectList()
                 .doOnSuccess(customers::addAll)
                 .doOnError(e -> log.error(e.getMessage(), e))
                 .subscribe(customers -> log.info("loaded {} customers", customers.size()));
+    }
+
+    @Async
+    @EventListener
+    public void handlePropertiesChanged(PropertiesChangedEvent event) {
+        globalProperties = event.getSource();
+        cacheStateService.markDirty("CACHE", true);
+        if (!reload())
+            refresh();
+    }
+
+    @Async
+    @EventListener
+    public void handleCustomersReloaded(CustomersReloadedEvent event) {
+        refresh();
     }
